requirements_test.txt
# Test requirements for Cat Care Tracker
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
